perf_debug: zone duration widens to uint64
A uint32 duration saturates at ~3 s of device cycles, which real zones
can exceed. The union member was already padded to 16 B by the u64
start's alignment, so the record stays 32 B and the widening is free;
the saturation clamp and its teardown counter go away with the reason
for them.

Verified on bh-05: pairing clean (13,220 delivered, 0 dropped, no
pairing warnings), ops-CSV pytest passes.

+++ tt_metal/tools/profiler/perf_debug_consumer.hpp
@@ -64,7 +64,7 @@ struct PerfDebugRec {
         uint64_t payload;  // Cont: one payload uint64 (hi word first on the wire)
         struct {
             uint64_t start;     // device timestamp of the zone open
-            uint32_t duration;  // device cycles; saturates at UINT32_MAX (~3 s @ 1.35 GHz, counted)
+            uint64_t duration;  // device cycles
         } zone;                 // Zone
     } data;
     uint32_t id;  // full 27-bit structural zone id (tu_id << TT_ZONE_LOCAL_BITS | local)
